move byte array formatting to the parquetgridview instead
